requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pillow==10.4.0
lxml==5.2.0
cssutils==2.11.1
aiofiles==24.1.0
jinja2==3.1.4
